requirements.txt
google-cloud-aiplatform[agent_engines,adk]>=1.55.0
google-adk>=0.1.0
langchain>=0.2.0
langchain-core>=0.2.0
python-dotenv>=1.0.0
google-auth>=2.29.0
google-genai>=0.5.0